+++ NEWS.markdown
@@ -1,6 +1,13 @@
 Revision history for jslint4java
 ================================
 
+2.0.4 (2013-04-10)
+------------------
+
+* Update to JSLint 2013-04-06.
+  - This removes the `anon`, `css`, `fragment` and `on` options.
+  - This adds the `closure` option.
+
 2.0.3 (2012-12-04)
 ------------------